pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "xml-lib"
version = "0.1.0"
description = "Production-grade XML lifecycle validator, guardrail subsystem, and mathematical engine"
authors = ["XML-Lib Contributors"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/farukalpay/xml-lib"
repository = "https://github.com/farukalpay/xml-lib"
documentation = "https://farukalpay.github.io/xml-lib"
keywords = ["xml", "validation", "lifecycle", "guardrails", "hilbert", "banach"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Text Processing :: Markup :: XML",
]
packages = [{include = "xml_lib"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
lxml = "^5.0.0"
xmlschema = "^2.5.0"
sympy = "^1.12"
numpy = "^1.26.0"
python-pptx = "^0.6.23"
cryptography = "^41.0.0"
jsonlines = "^4.0.0"
pyyaml = "^6.0"
jinja2 = "^3.1.0"
psycopg2-binary = "^2.9.0"
networkx = "^3.2.0"
graphviz = "^0.20.0"
plotly = "^5.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-benchmark = "^4.0.0"
hypothesis = "^6.92.0"
mypy = "^1.7.0"
ruff = "^0.1.0"
black = "^23.0.0"
pre-commit = "^3.5.0"
mkdocs-material = "^9.5.0"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.0"
mkdocs-section-index = "^0.3.0"
types-pyyaml = "^6.0.0"
types-jsonlines = "^4.0.0"
z3-solver = "^4.12.0"

[tool.poetry.scripts]
xml-lib = "xml_lib.cli_new:app"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/**/*.py" = ["ARG", "S101"]  # allow assert and unused args in tests

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "lxml.*",
    "xmlschema.*",
    "pptx.*",
    "graphviz.*",
    "plotly.*",
    "z3.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=xml_lib",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "property: marks property-based tests",
]

[tool.coverage.run]
branch = true
source = ["xml_lib"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
